pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spatial-reasoning"
version = "0.1.8"
description = "A PyPI package for object detection using advanced vision models"
readme = "README.md"
authors = [
    { name = "Qasim Wani", email = "qasim31wani@gmail.com" }
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
keywords = ["computer vision", "object detection", "AI", "machine learning", "OpenAI", "Gemini"]
dependencies = [
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=9.0.0",
    "opencv-python>=4.5.0",
    "matplotlib>=3.3.0",
    "scipy>=1.7.0",
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "openai>=1.0.0",
    "google-genai",
    "transformers>=4.30.0",
    "datasets>=2.0.0",
    "accelerate>=1.6.0",
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/QasimWani/spatial-reasoning"
"Bug Tracker" = "https://github.com/QasimWani/spatial-reasoning/issues"
Documentation = "https://github.com/QasimWani/spatial-reasoning#readme"
"Source Code" = "https://github.com/QasimWani/spatial-reasoning"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.990",
]

[project.scripts]
spatial-reasoning = "spatial_reasoning.run_cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["**/*.json", "**/*.txt", "**/*.md"]